Renaming of package 'dvn' to 'dataverse_connect' and initial update to handle private layers
